fixed bio deleting image

+++ api/controllers/bio.js
@@ -2,6 +2,7 @@ import {db} from '../db.js';
 import jwt from 'jsonwebtoken';
 import DOMPurify from 'isomorphic-dompurify'
 import dotenv from 'dotenv';
+import {v2 as cloudinary} from 'cloudinary'
 
 dotenv.config();
 
@@ -25,25 +26,56 @@ export const getBio =  (req, res, next) => {
     });
 };
 
+cloudinary.config({
+  cloud_name: process.env.REACT_APP_CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.REACT_APP_CLOUDINARY_KEY,
+  api_secret: process.env.REACT_APP_CLOUDINARY_SECRET_KEY,
+  secure : true
+});
+
 //effacer un album par son id avec comparaison du token
 export const deleteBio = (req, res, next) => {
-    const token = req.cookies.access_token;
-    if (!token) return res.status(401).json("Vous devez être connecté pour supprimer un texte bio.");
-  
-    jwt.verify(token, process.env.JWT_SECRET, (err) => {
-        if (err) return res.status(403).json("Vous n'êtes pas autorisé à supprimer ce texte bio.");
+  const token = req.cookies.access_token;
+  if (!token) return res.status(401).json("Vous devez être connecté pour supprimer un texte bio.");
 
-        const bioId = req.params.id;
+  jwt.verify(token, process.env.JWT_SECRET, (err) => {
+    if (err) return res.status(403).json("Vous n'êtes pas autorisé à supprimer ce texte bio.");
 
-        const deleteQuery = 'DELETE FROM bios WHERE id = ?';
-        db.query(deleteQuery, [bioId], (err) => {
+    const bioId = req.params.id;
+    const q = 'SELECT image FROM bios WHERE id = ?';
+
+    db.query(q, [bioId], (err, data) => {
+      if (err) {
+        next(err);
+        return;
+      }
+
+      const bio = data[0];
+      if (!bio) {
+        return res.status(404).json("Le texte bio n'a pas été trouvé.");
+      }
+
+      const imageURL = bio.image;
+      const imagePublicId = imageURL.split('/').pop().split('.')[0];
+
+      // Efface l'image de cloudinary
+      cloudinary.uploader.destroy(imagePublicId)
+        .then(() => {
+          const deleteQuery = 'DELETE FROM bios WHERE id = ?';
+          db.query(deleteQuery, [bioId], (err) => {
             if (err) {
-                next(err);
-                return;
+              next(err);
+              return;
             }
-            return res.status(200).json('le texte a été supprimée avec succès.');
+            return res.status(200).json('Le texte bio et son image ont été supprimés avec succès.');
+          });
+        })
+        .catch((error) => {
+          console.error('Erreur lors de la suppression de l\'image sur Cloudinary:', error);
+          return res.status(500).json('Une erreur s\'est produite lors de la suppression de l\'image.');
         });
     });
+  });
 };
 
 // ajouter un album
